Critical Fix
Started add the loop to create and execute in the Command Design Pattern
directory.  I started an implementation, but decided to make each command
have the same return type.

After fix the return did a lot of refactoring.

The next step is to circle back to implementing the "execution of
commands" command.